using browserify, image optimization
