+++ kernels/index.html
@@ -128,9 +128,7 @@
 
   <script src="../bower_components/jquery/dist/jquery.min.js"></script>
   <script src="../bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
-  <script src="../bower_components/d3/d3.min.js"></script>
-  <script src="../bower_components/moment/moment.js"></script>
-  <script src="../scripts/kernels.js"></script>
+  <script src="./built.js"></script>
 
   <script type="text/javascript" src="https://use.typekit.net/kpd5yrh.js"></script>
   <script type="text/javascript">try{Typekit.load();}catch(e){}</script>
